Update main.ipynb
Tweaked the verify() method in the MiminalChain class as the backdating check's if clause comparison of block timestamps was reversed, rendering a result that the current block's timestamp was backdated compared to the previous block's timestamp, when in reality it was not.
Added the ForkingEnums class so that the fork() method's head parameter would be a consistent integer, rather than a mixture of string and integer.
Reduced the number of test blocks generated to 10 instead of 20, adjusting hardcoded indexes to accommodate this change as needed. 
Add a __repr__ method to the MinimalBlock class to facilitate easy printing of the block, useful when one wishes to print a block to the console.